ETCM-168: Moved service building to a separate file.
